Fixed author error, hopefully fixing empty data error
Added an "if the author is there" check in line 117 of
get_fb_posts_fb_group.py to stop the KeyError that was happenning --
this seems to work.
Uncommented a line to see whether the lack of it was preventing writing
anything other than the headers in the spreadsheet.
Hopefully it works!